Add clones subdir and script
A new feature. Handy powershell repositories now go in the Clones
folder. Similar to Modules, these scripts do not sync to the Lib.PS
repo, but can be cloned and integrated into the lib.

I also added the reset-clones.ps1 script to simplify the init (and
re-init) of the library's clones!