Added Try/Catch to Get-SplitEnvPath
Found an instance where 'Access Denied' was hit when test-path`ing a
dir. Figured I'd add a catch to put the error in the 'Exists' column.

Also, might have found a glitch in the reformat the VS Code PowerShell
module does. Not sure why, but it's beats the heck out of the format for
Get-Profiles. Even if I re-jigger it with better spaces and no
backticks....